+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.ht
 
 ## [Unreleased]
 
+### Changed
+
+- Reduced the header height slightly to provide more room for content.
+
 ## [0.2.21] - 2026-07-28
 
 ### Added